+++ src/forms/userAccount.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { withFormik, Form, Field } from "formik";
 import * as Yup from "yup";
+import "../../src/scss/userAccount.scss";
 
 function UserAccountForm({ values, errors, touched, isSubmitting }) {
   return (
@@ -41,21 +42,6 @@ const FormikUserAccountForm = withFormik({
   handleSubmit(values, { resetForm, setErrors, setSubmitting }) {
       console.log("User account form saved changes");
       console.log(values);
-    // if (values.email === "alreadytaken@atb.dev") {
-    //   setErrors({ email: "That email is already taken" });
-    // } else {
-    //   axios
-    //     .post("https://yourdatabaseurlgoeshere.com", values)
-    //     .then(res => {
-    //       console.log(res); // Data was created successfully and logs to console
-    //       resetForm();
-    //       setSubmitting(false);
-    //     })
-    //     .catch(err => {
-    //       console.log(err); // There was an error creating the data and logs to console
-    //       setSubmitting(false);
-    //     });
-    // }
   }
 })(UserAccountForm);
